[ticket] refactor: useOAuth 훅 리팩토링 - 상수 및 타입 분리, 소셜 로그인 로직 개선
