+++ apps/ticket/src/shared/hooks/useOAuth/constants.ts
@@ -1,0 +1,4 @@
+export const KAKAO_REST_API_KEY = process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY;
+export const GOOGLE_CLIENT_ID = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID;
+
+export const REDIRECT_URI = process.env.NEXT_PUBLIC_REDIRECT_URI;
