+++ apps/ticket/src/shared/hooks/useOAuth/types.ts
@@ -1,0 +1,8 @@
+import { ObjectValues } from "@/shared/types/utils";
+
+export const SOCIAL_LOGIN_TYPE = {
+  KAKAO: "KAKAO",
+  GOOGLE: "GOOGLE",
+} as const;
+
+export type SocialLoginType = ObjectValues<typeof SOCIAL_LOGIN_TYPE>;
